changed benches to use seeded rng
