+++ yrs/benches/benches.rs
@@ -1,136 +1,161 @@
 use criterion::*;
-use lib0::decoding::Cursor;
 use rand::distributions::Alphanumeric;
-use rand::prelude::ThreadRng;
-use rand::{thread_rng, Rng};
+use rand::prelude::StdRng;
+use rand::{Rng, RngCore, SeedableRng};
 use std::cell::Cell;
-use std::fs::File;
-use std::io::Read;
-use yrs::updates::decoder::{Decode, Decoder, DecoderV1};
-use yrs::updates::encoder::{Encode, Encoder};
-use yrs::{Array, Doc, Text};
+use yrs::Doc;
 
-#[derive(PartialEq, Eq, Debug)]
+const N: usize = 6000;
+const SEED: u64 = 0xdeadbeaf;
+
 enum TextOp {
     Insert(u32, String),
     Delete(u32, u32),
 }
 
-impl Encode for TextOp {
-    fn encode<E: Encoder>(&self, encoder: &mut E) {
-        match self {
-            TextOp::Insert(idx, content) => {
-                encoder.write_u32(1);
-                encoder.write_u32(*idx);
-                encoder.write_string(content.as_str());
-            }
-            TextOp::Delete(idx, len) => {
-                encoder.write_u32(2);
-                encoder.write_u32(*idx);
-                encoder.write_u32(*len);
-            }
-        }
-    }
-}
-
-impl Decode for TextOp {
-    fn decode<D: Decoder>(decoder: &mut D) -> Self {
-        match decoder.read_u32() {
-            1 => {
-                let idx = decoder.read_u32();
-                let content = decoder.read_string().to_string();
-                TextOp::Insert(idx, content)
-            }
-            2 => {
-                let idx = decoder.read_u32();
-                let len = decoder.read_u32();
-                TextOp::Delete(idx, len)
-            }
-            a => panic!("unrecognized TextOp case: {}", a),
-        }
-    }
-}
-
-#[derive(PartialEq, Eq, Debug)]
 enum ArrayOp {
     Insert(u32, Vec<u32>),
     Delete(u32, u32),
 }
 
-impl Encode for ArrayOp {
-    fn encode<E: Encoder>(&self, encoder: &mut E) {
-        match self {
-            ArrayOp::Insert(idx, content) => {
-                encoder.write_u32(1);
-                encoder.write_u32(*idx);
-                encoder.write_u32(content.len() as u32);
-                for i in content.iter() {
-                    encoder.write_u32(*i);
-                }
+fn b1_1<R: RngCore>(rng: &mut R, size: usize) -> Vec<TextOp> {
+    let sample: Vec<_> = rng
+        .sample_iter(&Alphanumeric)
+        .take(size)
+        .map(|c| c.to_string())
+        .collect();
+
+    (0..size as u32)
+        .into_iter()
+        .zip(sample)
+        .map(|(i, str)| TextOp::Insert(i, str))
+        .collect()
+}
+
+fn b1_2<R: RngCore>(rng: &mut R, size: usize) -> Vec<TextOp> {
+    let s: String = rng.sample_iter(&Alphanumeric).take(size as usize).collect();
+    vec![TextOp::Insert(0, s)]
+}
+
+fn b1_3<R: RngCore>(rng: &mut R, size: usize) -> Vec<TextOp> {
+    let sample: Vec<_> = rng
+        .sample_iter(&Alphanumeric)
+        .take(size)
+        .map(|c| c.to_string())
+        .collect();
+
+    (0..size as u32)
+        .into_iter()
+        .zip(sample)
+        .map(|(i, str)| TextOp::Insert(0, str))
+        .collect()
+}
+
+fn b1_4<R: RngCore>(rng: &mut R, size: usize) -> Vec<TextOp> {
+    let sample: Vec<_> = rng
+        .sample_iter(&Alphanumeric)
+        .take(size)
+        .map(|c| c.to_string())
+        .collect();
+
+    (0..size as u32)
+        .into_iter()
+        .zip(sample)
+        .map(|(i, str)| {
+            let idx = rng.gen_range(0, i.max(1));
+            TextOp::Insert(idx, str)
+        })
+        .collect()
+}
+
+fn gen_string<R: RngCore>(rng: &mut R, min: usize, max: usize) -> String {
+    let len = rng.gen_range(min, max);
+    rng.sample_iter(&Alphanumeric).take(len).collect()
+}
+
+fn b1_5<R: RngCore>(rng: &mut R, size: usize) -> Vec<TextOp> {
+    (0..size as u32)
+        .into_iter()
+        .map(|i| {
+            let str = gen_string(rng, 2, 10);
+            let idx = rng.gen_range(0, i.max(1));
+            TextOp::Insert(idx, str)
+        })
+        .collect()
+}
+
+fn b1_6<R: RngCore>(rng: &mut R, size: usize) -> Vec<TextOp> {
+    let s: String = rng.sample_iter(&Alphanumeric).take(size as usize).collect();
+    let len = s.len() as u32;
+    vec![TextOp::Insert(0, s), TextOp::Delete(0, len)]
+}
+
+fn b1_7<R: RngCore>(rng: &mut R, size: usize) -> Vec<TextOp> {
+    let mut total_len = Cell::new(0u32);
+    (0..size as u32)
+        .into_iter()
+        .map(|i| {
+            let total = total_len.get();
+            let idx = if total == 0 {
+                0
+            } else {
+                rng.gen_range(0, total)
+            };
+            if total == idx || rng.gen_bool(0.5) {
+                let str = {
+                    let len = rng.gen_range(2, 10);
+                    total_len.set(total + len);
+                    rng.sample_iter(&Alphanumeric).take(len as usize).collect()
+                };
+                TextOp::Insert(idx, str)
+            } else {
+                let hi = (total - idx).min(9);
+                let len = if hi == 1 { 1 } else { rng.gen_range(1, hi) };
+                total_len.set(total - len);
+                TextOp::Delete(idx, len)
             }
-            ArrayOp::Delete(idx, len) => {
-                encoder.write_u32(2);
-                encoder.write_u32(*idx);
-                encoder.write_u32(*len);
-            }
-        }
-    }
+        })
+        .collect()
 }
 
-impl Decode for ArrayOp {
-    fn decode<D: Decoder>(decoder: &mut D) -> Self {
-        match decoder.read_u32() {
-            1 => {
-                let idx = decoder.read_u32();
-                let mut i = decoder.read_u32() as usize;
-                let mut content = Vec::with_capacity(i);
-                while i > 0 {
-                    content.push(decoder.read_u32());
-                    i -= 1;
-                }
-                ArrayOp::Insert(idx, content)
-            }
-            2 => {
-                let idx = decoder.read_u32();
-                let len = decoder.read_u32();
-                ArrayOp::Delete(idx, len)
-            }
-            _ => panic!("unrecognized TextOp case"),
-        }
-    }
+fn b1_8<R: RngCore>(rng: &mut R, size: usize) -> Vec<ArrayOp> {
+    let ops: Vec<ArrayOp> = (0..size)
+        .map(|i| ArrayOp::Insert(i as u32, vec![rng.gen()]))
+        .collect();
+    ops
 }
 
-fn encode_vec<T: Encode, E: Encoder>(vec: &Vec<T>, encoder: &mut E) {
-    encoder.write_u32(vec.len() as u32);
-    for value in vec.iter() {
-        value.encode(encoder);
-    }
+fn b1_9<R: RngCore>(rng: &mut R, size: usize) -> Vec<ArrayOp> {
+    let sample: Vec<u32> = (0..size).map(|_| rng.gen()).collect();
+    vec![ArrayOp::Insert(0, sample)]
 }
 
-fn decode_vec<T: Decode, D: Decoder>(decoder: &mut D) -> Vec<T> {
-    let mut i = decoder.read_u32() as usize;
-    let mut vec = Vec::with_capacity(i);
-    while i > 0 {
-        vec.push(T::decode(decoder));
-        i -= 1;
-    }
-    vec
+fn b1_10<R: RngCore>(rng: &mut R, size: usize) -> Vec<ArrayOp> {
+    (0..size)
+        .map(|i| ArrayOp::Insert(0, vec![rng.gen()]))
+        .collect()
 }
 
-fn load<T: Decode>(src: &str) -> Vec<T> {
-    let mut f = File::open(src).unwrap();
-    let mut buf = Vec::new();
-    f.read_to_end(&mut buf).unwrap();
-    let mut decoder = DecoderV1::new(Cursor::new(buf.as_slice()));
-    decode_vec(&mut decoder)
+fn b1_11<R: RngCore>(rng: &mut R, size: usize) -> Vec<ArrayOp> {
+    (0..size)
+        .map(|i| {
+            let idx = rng.gen_range(0, (i as u32).max(1));
+            let values = vec![rng.gen()];
+            ArrayOp::Insert(idx, values)
+        })
+        .collect()
 }
 
-fn text_benchmark(c: &mut Criterion, name: &str, src: &str) {
+fn text_benchmark<F>(c: &mut Criterion, name: &str, gen: F)
+where
+    F: FnOnce(&mut StdRng, usize) -> Vec<TextOp>,
+{
     let input = {
         let doc = Doc::new();
         let mut txn = doc.transact();
         let txt = txn.get_text("text");
-        let ops = load(src);
+        let mut rng = StdRng::seed_from_u64(SEED);
+        let ops = gen(&mut rng, N);
         (doc, txt, ops)
     };
 
@@ -151,12 +176,16 @@ fn text_benchmark(c: &mut Criterion, name: &str, src: &str) {
     );
 }
 
-fn array_benchmark(c: &mut Criterion, name: &str, src: &str) {
+fn array_benchmark<F>(c: &mut Criterion, name: &str, gen: F)
+where
+    F: FnOnce(&mut StdRng, usize) -> Vec<ArrayOp>,
+{
     let input = {
         let doc = Doc::new();
         let mut txn = doc.transact();
         let array = txn.get_array("text");
-        let ops = load(src);
+        let mut rng = StdRng::seed_from_u64(SEED);
+        let ops = gen(&mut rng, N);
         (doc, array, ops)
     };
 
@@ -179,106 +208,18 @@ fn array_benchmark(c: &mut Criterion, name: &str, src: &str) {
     );
 }
 
-fn b1_1(c: &mut Criterion) {
-    text_benchmark(
-        c,
-        "[B1.1] Append 6000 characters",
-        "./yrs/benches/input/b-1-1-6000.bin",
-    );
-}
-
-fn b1_2(c: &mut Criterion) {
-    text_benchmark(
-        c,
-        "[B1.2] Insert string of length 6000",
-        "./yrs/benches/input/b-1-2-6000.bin",
-    );
-}
-
-fn b1_3(c: &mut Criterion) {
-    text_benchmark(
-        c,
-        "[B1.3] Prepend 6000 characters",
-        "./yrs/benches/input/b-1-3-6000.bin",
-    );
-}
-
-fn b1_4(c: &mut Criterion) {
-    text_benchmark(
-        c,
-        "[B1.4] Insert 6000 characters at random positions",
-        "./yrs/benches/input/b-1-4-6000.bin",
-    );
-}
-
-fn b1_5(c: &mut Criterion) {
-    text_benchmark(
-        c,
-        "[B1.5] Insert 6000 words at random positions",
-        "./yrs/benches/input/b-1-5-6000.bin",
-    );
-}
-
-fn b1_6(c: &mut Criterion) {
-    text_benchmark(
-        c,
-        "[B1.6] Insert string, then delete it",
-        "./yrs/benches/input/b-1-6-6000.bin",
-    );
-}
-
-fn b1_7(c: &mut Criterion) {
-    text_benchmark(
-        c,
-        "[B1.7] Insert/Delete strings at random positions",
-        "./yrs/benches/input/b-1-7-6000.bin",
-    );
-}
-
-fn b1_8(c: &mut Criterion) {
-    array_benchmark(
-        c,
-        "[B1.8] Append 6000 numbers",
-        "./yrs/benches/input/b-1-8-6000.bin",
-    );
-}
-
-fn b1_9(c: &mut Criterion) {
-    array_benchmark(
-        c,
-        "[B1.9] Insert Array of 6000 numbers",
-        "./yrs/benches/input/b-1-9-6000.bin",
-    );
-}
-
-fn b1_10(c: &mut Criterion) {
-    array_benchmark(
-        c,
-        "[B1.10] Prepend 6000 numbers",
-        "./yrs/benches/input/b-1-10-6000.bin",
-    );
-}
-
-fn b1_11(c: &mut Criterion) {
-    array_benchmark(
-        c,
-        "[B1.11] Insert 6000 numbers at random positions",
-        "./yrs/benches/input/b-1-11-6000.bin",
-    );
-}
-
 fn bench(c: &mut Criterion) {
-    b1_1(c);
-    b1_2(c);
-    b1_3(c);
-    b1_4(c);
-    b1_5(c);
-    b1_6(c);
-    b1_7(c);
-    b1_8(c);
-    b1_9(c);
-    b1_10(c);
-    b1_11(c);
+    text_benchmark(c, "[B1.1] Append N characters", b1_1);
+    text_benchmark(c, "[B1.2] Insert string of length N", b1_2);
+    text_benchmark(c, "[B1.3] Prepend N characters", b1_3);
+    text_benchmark(c, "[B1.4] Insert N characters at random positions", b1_4);
+    text_benchmark(c, "[B1.5] Insert N words at random positions", b1_5);
+    text_benchmark(c, "[B1.6] Insert string, then delete it", b1_6);
+    text_benchmark(c, "[B1.7] Insert/Delete strings at random positions", b1_7);
+    array_benchmark(c, "[B1.8] Append N numbers", b1_8);
+    array_benchmark(c, "[B1.9] Insert Array of N numbers", b1_9);
+    array_benchmark(c, "[B1.10] Prepend N numbers", b1_10);
+    array_benchmark(c, "[B1.11] Insert N numbers at random positions", b1_11);
 }
 
 criterion_group! {
